refactor(app): extract DeviceManager to runtime/devices.py (#70)
Move the 13-method device-management cluster out of Runtime into a
collaborator class DeviceManager in src/vanchor/runtime/devices.py.
Runtime retains one-line shims on all 13 methods (external callers
in server.py, tests, and restore_backup keep working unchanged).
DeviceManager holds a self._rt back-reference for shared state.
